Command checking...
Added commands that can be written both by clients and by admins. Everyclient should have an unique id that you can recieve by connecting to a server for the first time. I also added a prefix so that you can see that you are talking with an admin.

Messages that start with the character "/" will not be displayed for everyone now.

The console responds with a custom prefix to it's message.