Settings.json added! :bulb:
Added settings.json. In there you will be able to set lots of stuff. With that being I also added some more cool stuff. Here it is :+1:

Timestamp: You can now chose if you want timestamp on or off. This was added in this version...

Versioncontrol: When starting up the script it will now show you what version you are running and if you are up to the correct version.. This can be toggled inside the settings.json file...

Settings.json default:

{
    "timestamp": true,
    "adminprefix": true,
    "colors": true,
    "showUpdates": true
}

Alright, see you in the next update post! :+1: